Move execa mock under packages tests
