--- __mocks__/execa.js
@@ -1,1 +1,0 @@
-module.exports = jest.genMockFromModule("execa");
