+++ packages/core-cli/__tests__/__mocks__/execa.js
@@ -1,0 +1,1 @@
+module.exports = jest.createMockFromModule("execa");
