+++ packages/core-cli/__tests__/services/installer.test.ts
@@ -1,10 +1,10 @@
 import "jest-extended";
 
-import { Console } from "@packages/core-test-framework";
 import { Installer } from "@packages/core-cli/src/services";
+import { Console } from "@packages/core-test-framework";
 import { setGracefulCleanup } from "tmp";
 
-import execa from "../../../../__mocks__/execa";
+import execa from "../__mocks__/execa";
 
 let cli;
 let installer;
