+++ packages/core/__tests__/commands/config-cli.test.ts
@@ -1,8 +1,8 @@
+import { Command } from "@packages/core/src/commands/config-cli";
 import { Container } from "@packages/core-cli";
 import { Console } from "@packages/core-test-framework";
-import { Command } from "@packages/core/src/commands/config-cli";
 
-import execa from "../../../../__mocks__/execa";
+import execa from "../__mocks__/execa";
 
 let cli;
 let config;
