+++ packages/core/__tests__/commands/reinstall.test.ts
@@ -1,11 +1,11 @@
 import "jest-extended";
 
+import { Command } from "@packages/core/src/commands/reinstall";
 import { Container } from "@packages/core-cli";
 import { Console } from "@packages/core-test-framework";
-import { Command } from "@packages/core/src/commands/reinstall";
 import prompts from "prompts";
 
-import execa from "../../../../__mocks__/execa";
+import execa from "../__mocks__/execa";
 
 let cli;
 let processManager;
