+++ packages/core/__tests__/commands/update.test.ts
@@ -1,12 +1,12 @@
 import "jest-extended";
 
+import { Command } from "@packages/core/src/commands/update";
 import { Container } from "@packages/core-cli";
 import { Console } from "@packages/core-test-framework";
-import { Command } from "@packages/core/src/commands/update";
 import nock from "nock";
 import prompts from "prompts";
 
-import execa from "../../../../__mocks__/execa";
+import execa from "../__mocks__/execa";
 import { versionNext } from "../internal/__fixtures__/latest-version";
 
 let cli;
